Fix broken links in automatic help index
com_help generates an index automatically when a component doesn't provide one.  However, the links in the index were broken.   That's because com_help removes the file extension, and re-adds it when links are accessed, but the file extension information (layoutExt) was missing at the removal step, so it only removed a dot.